PaymentSession: Use CompletableFuture internally (rather than ListenableCompletableFuture)
Also in support of / as a result of this change:

* remove supplyAsync from ListenableCompletableFuture
* move ThrowingSupplier to FutureUtils
