+++ core/src/main/java/org/bitcoinj/utils/ListenableCompletableFuture.java
@@ -86,52 +86,4 @@ public class ListenableCompletableFuture<V> extends CompletableFuture<V> impleme
         }
         return listenable;
     }
-
-    public static <T> ListenableCompletableFuture<T> supplyAsync(ThrowingSupplier<T> throwingSupplier, Executor executor) {
-        ListenableCompletableFuture<T> future = new ListenableCompletableFuture<>();
-        executor.execute(() -> {
-            try {
-                T result = throwingSupplier.getThrows();
-                future.complete(result);
-            } catch (Exception e) {
-                future.completeExceptionally(e);
-            }
-        });
-        return future;
-    }
-
-    /**
-     * Subinterface of {@link Supplier} for Lambdas which throw exceptions.
-     * Can be used for two purposes:
-     * 1. To cast a lambda that throws an exception to a {@link Supplier} and
-     * automatically wrapping any exceptions with {@link RuntimeException}.
-     * 2. As a {@code FunctionalInterface} where a lambda that throws exceptions is
-     * expected or allowed.
-     *
-     * @param <T>
-     */
-    @FunctionalInterface
-    public interface ThrowingSupplier<T> extends Supplier<T> {
-
-        /**
-         * Gets a result wrapping checked Exceptions with {@link RuntimeException}
-         * @return a result
-         */
-        @Override
-        default T get() {
-            try {
-                return getThrows();
-            } catch (final Exception e) {
-                throw new RuntimeException(e);
-            }
-        }
-
-        /**
-         * Gets a result.
-         *
-         * @return a result
-         * @throws Exception Any checked Exception
-         */
-        T getThrows() throws Exception;
-    }
 }
